Shift out hardcoded editor and add more security to /api/open
Now that any editor exe can be specified in the .env file, it is made sure that only a valid file inside the todo folder can be passed as argument. That theoretically does not stop anyone from e.g. specifying python as their editor and putting a python file in their todo folder, but at least this way it's their own fault.